added form component, and parent to handle setting bulk allocations
